Update project structre

+++ short_octal_format.c
@@ -16,7 +16,7 @@ int short_octal(va_list ptr, char *buf, int *buf_ind, char *flag, int w)
 {
 	unsigned short i = va_arg(ptr, int);
 	unsigned short tmp = i;
-	int divider = 1, printed_characters = 0;
+	unsigned short divider = 1, printed_characters = 0;
 	char num;
 	int x, len = 0, remind;
